+++ src/test/csfle/spec.rs
@@ -1,4 +1,4 @@
-use crate::test::spec::{unified_runner::run_unified_tests, v2_runner::run_v2_tests};
+use crate::test::spec::unified_runner::run_unified_tests;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn run_unified() {
@@ -17,26 +17,5 @@ async fn run_unified() {
         .skip_tests(&skipped_tests)
         // TODO RUST-582: unskip this test
         .skip_files(&["timeoutMS.json"])
-        .await;
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn run_legacy() {
-    let mut skipped_files = vec![
-        // TODO RUST-528: unskip this file
-        "timeoutMS.json",
-        // These files have been migrated to unified tests.
-        // TODO DRIVERS-3178 remove these once the files are gone.
-        "fle2v2-BypassQueryAnalysis.json",
-        "fle2v2-EncryptedFields-vs-EncryptedFieldsMap.json",
-        "localSchema.json",
-        "maxWireVersion.json",
-    ];
-    if cfg!(not(feature = "openssl-tls")) {
-        skipped_files.push("kmipKMS.json");
-    }
-
-    run_v2_tests(&["client-side-encryption", "legacy"])
-        .skip_files(&skipped_files)
         .await;
 }
